# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TrustHood Escrow API
  version: 0.1.0
  description: |
    REST API for the TrustHood Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints return a standardised envelope:
    ```
    {
      "data": [...],
      "page": 1, "limit": 20, "total": 42,
      "totalPages": 3, "hasNextPage": true, "hasPreviousPage": false
    }
    ```

    ## Environments
    - **Testnet**: `https://api.test.trusthood-escrow.example.com`
    - **Mainnet**: `https://api.trusthood-escrow.example.com`
  contact:
    name: TrustHood Escrow
    url: https://trusthood-escrow.example.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  x-repository: https://github.com/trusthood-escrow/trusthood-escrow
  x-docs-url: https://docs.trusthood-escrow.example.com
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.test.trusthood-escrow.example.com
    description: Testnet
  - url: https://api.trusthood-escrow.example.com
    description: Mainnet
...
security:
  - BearerAuth: []
paths:
...
  /api/users/{address}/export:
    get:
      tags: [Users]
      summary: Export all user data as JSON
      operationId: exportUserData
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Export data envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserExport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Address mismatch with authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{address}/import:
    post:
      tags: [Users]
      summary: Import user data from JSON
      operationId: importUserData
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  description: Previously exported user data
                mode:
                  type: string
                  enum: [merge, replace]
                  default: merge
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  results:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/users/{address}/export/file:
    get:
      tags: [Users]
      summary: Download user data as a JSON file
      operationId: downloadExportFile
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: JSON file download
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    type: string
                    example: data.json
                  content:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Address mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
...
    UserExport:
      type: object
      properties:
        version:
          type: string
          example: '1.0'
        exportedAt:
          type: string
          format: date-time
        userAddress:
          type: string
        data:
          type: object
          properties:
            escrows:
              type: array
            payments:
              type: array
            kyc:
              type: object
            reputation:
              type: object
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...